# libs/shared/apis/v1/synaptiq-apis/src/main/java/com/synaptiq/infrastructure/in/web/api/CatalogApi.java
/*
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.21.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.synaptiq.infrastructure.in.web.api;

import com.synaptiq.infrastructure.in.web.dto.CatalogImportResponse;
import com.synaptiq.infrastructure.in.web.dto.CatalogItemListResponse;
import com.synaptiq.infrastructure.in.web.dto.CatalogItemResponse;
import com.synaptiq.infrastructure.in.web.dto.CatalogSchemaResponse;
import com.synaptiq.infrastructure.in.web.dto.CreateCatalogItemRequest;
import java.util.Map;
import org.springframework.lang.Nullable;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import io.swagger.v3.oas.annotations.media.ExampleObject;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.multipart.MultipartFile;
import org.springframework.web.server.ServerWebExchange;
import reactor.core.publisher.Flux;
import reactor.core.publisher.Mono;
import org.springframework.http.codec.multipart.Part;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2026-05-05T21:15:25.464614100-05:00[America/Chicago]", comments = "Generator version: 7.21.0")
@Validated
@Tag(name = "Catalog", description = "Catalog Registry — schema management, item CRUD, embedding")
public interface CatalogApi {

    String PATH_CREATE_CATALOG_ITEM = "/api/v1/catalog/items";
    /**
     * POST /api/v1/catalog/items : Create a catalog item
     *
     * @param createCatalogItemRequest  (required)
     * @param xTenantID The tenant identifier for multi-tenant isolation (optional)
     * @return Item created (status code 201)
     */
    @Operation(
        operationId = "createCatalogItem",
        summary = "Create a catalog item",
        tags = { "Catalog" },
        responses = {
            @ApiResponse(responseCode = "201", description = "Item created", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = CatalogItemResponse.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = CatalogApi.PATH_CREATE_CATALOG_ITEM,
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    Mono<ResponseEntity<CatalogItemResponse>> createCatalogItem(
        @Parameter(name = "CreateCatalogItemRequest", description = "", required = true) @Valid @RequestBody Mono<CreateCatalogItemRequest> createCatalogItemRequest,
        @Size(min = 1, max = 100) @Parameter(name = "X-Tenant-ID", description = "The tenant identifier for multi-tenant isolation", in = ParameterIn.HEADER) @RequestHeader(value = "X-Tenant-ID", required = false) @Nullable String xTenantID,
        @Parameter(hidden = true) final ServerWebExchange exchange
    );


    String PATH_DELETE_CATALOG_ITEM = "/api/v1/catalog/items/{itemId}";
    /**
     * DELETE /api/v1/catalog/items/{itemId} : Delete a catalog item
     *
     * @param itemId  (required)
     * @param xTenantID The tenant identifier for multi-tenant isolation (optional)
     * @return Item deleted (status code 204)
     */
    @Operation(
        operationId = "deleteCatalogItem",
        summary = "Delete a catalog item",
        tags = { "Catalog" },
        responses = {
            @ApiResponse(responseCode = "204", description = "Item deleted")
        }
    )
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = CatalogApi.PATH_DELETE_CATALOG_ITEM
    )
    Mono<ResponseEntity<Void>> deleteCatalogItem(
        @Parameter(name = "itemId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("itemId") String itemId,
        @Size(min = 1, max = 100) @Parameter(name = "X-Tenant-ID", description = "The tenant identifier for multi-tenant isolation", in = ParameterIn.HEADER) @RequestHeader(value = "X-Tenant-ID", required = false) @Nullable String xTenantID,
        @Parameter(hidden = true) final ServerWebExchange exchange
    );


    String PATH_GET_ACTIVE_SCHEMA = "/api/v1/catalog/schema";
    /**
     * GET /api/v1/catalog/schema : Get active catalog schema
     *
     * @param xTenantID The tenant identifier for multi-tenant isolation (optional)
     * @return Active schema (status code 200)
     */
    @Operation(
        operationId = "getActiveSchema",
        summary = "Get active catalog schema",
        tags = { "Catalog" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Active schema", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = CatalogSchemaResponse.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = CatalogApi.PATH_GET_ACTIVE_SCHEMA,
        produces = { "application/json" }
    )
    Mono<ResponseEntity<CatalogSchemaResponse>> getActiveSchema(
        @Size(min = 1, max = 100) @Parameter(name = "X-Tenant-ID", description = "The tenant identifier for multi-tenant isolation", in = ParameterIn.HEADER) @RequestHeader(value = "X-Tenant-ID", required = false) @Nullable String xTenantID,
        @Parameter(hidden = true) final ServerWebExchange exchange
    );


    String PATH_GET_CATALOG_ITEM = "/api/v1/catalog/items/{itemId}";
    /**
     * GET /api/v1/catalog/items/{itemId} : Get a catalog item
     *
     * @param itemId  (required)
     * @param xTenantID The tenant identifier for multi-tenant isolation (optional)
     * @return Item details (status code 200)
     */
    @Operation(
        operationId = "getCatalogItem",
        summary = "Get a catalog item",
        tags = { "Catalog" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Item details", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = CatalogItemResponse.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = CatalogApi.PATH_GET_CATALOG_ITEM,
        produces = { "application/json" }
    )
    Mono<ResponseEntity<CatalogItemResponse>> getCatalogItem(
        @Parameter(name = "itemId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("itemId") String itemId,
        @Size(min = 1, max = 100) @Parameter(name = "X-Tenant-ID", description = "The tenant identifier for multi-tenant isolation", in = ParameterIn.HEADER) @RequestHeader(value = "X-Tenant-ID", required = false) @Nullable String xTenantID,
        @Parameter(hidden = true) final ServerWebExchange exchange
    );


    String PATH_IMPORT_CATALOG_ITEMS = "/api/v1/catalog/import";
    /**
     * POST /api/v1/catalog/import : Import catalog items from CSV
     *
     * @param xTenantID The tenant identifier for multi-tenant isolation (optional)
     * @param file  (optional)
     * @return Import results (status code 200)
     */
    @Operation(
        operationId = "importCatalogItems",
        summary = "Import catalog items from CSV",
        tags = { "Catalog" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Import results", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = CatalogImportResponse.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = CatalogApi.PATH_IMPORT_CATALOG_ITEMS,
        produces = { "application/json" },
        consumes = { "multipart/form-data" }
    )
    Mono<ResponseEntity<CatalogImportResponse>> importCatalogItems(
        @Size(min = 1, max = 100) @Parameter(name = "X-Tenant-ID", description = "The tenant identifier for multi-tenant isolation", in = ParameterIn.HEADER) @RequestHeader(value = "X-Tenant-ID", required = false) @Nullable String xTenantID,
        @Parameter(name = "file", description = "") @RequestPart(value = "file", required = false) Part file,
        @Parameter(hidden = true) final ServerWebExchange exchange
    );


    String PATH_LIST_CATALOG_ITEMS = "/api/v1/catalog/items";
    /**
     * GET /api/v1/catalog/items : List catalog items
     *
     * @param xTenantID The tenant identifier for multi-tenant isolation (optional)
     * @param statusFilter  (optional)
     * @param skip  (optional, default to 0)
     * @param limit  (optional, default to 20)
     * @return Item list (status code 200)
     */
    @Operation(
        operationId = "listCatalogItems",
        summary = "List catalog items",
        tags = { "Catalog" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Item list", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = CatalogItemListResponse.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = CatalogApi.PATH_LIST_CATALOG_ITEMS,
        produces = { "application/json" }
    )
    Mono<ResponseEntity<CatalogItemListResponse>> listCatalogItems(
        @Size(min = 1, max = 100) @Parameter(name = "X-Tenant-ID", description = "The tenant identifier for multi-tenant isolation", in = ParameterIn.HEADER) @RequestHeader(value = "X-Tenant-ID", required = false) @Nullable String xTenantID,
        @Parameter(name = "statusFilter", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "statusFilter", required = false) @Nullable String statusFilter,
        @Parameter(name = "skip", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "skip", required = false, defaultValue = "0") Integer skip,
        @Parameter(name = "limit", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "limit", required = false, defaultValue = "20") Integer limit,
        @Parameter(hidden = true) final ServerWebExchange exchange
    );


    String PATH_UPDATE_CATALOG_ITEM = "/api/v1/catalog/items/{itemId}";
    /**
     * PATCH /api/v1/catalog/items/{itemId} : Update a catalog item
     *
     * @param itemId  (required)
     * @param requestBody  (required)
     * @param xTenantID The tenant identifier for multi-tenant isolation (optional)
     * @return Item updated (status code 200)
     */
    @Operation(
        operationId = "updateCatalogItem",
        summary = "Update a catalog item",
        tags = { "Catalog" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Item updated", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = CatalogItemResponse.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.PATCH,
        value = CatalogApi.PATH_UPDATE_CATALOG_ITEM,
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    Mono<ResponseEntity<CatalogItemResponse>> updateCatalogItem(
        @Parameter(name = "itemId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("itemId") String itemId,
        @Parameter(name = "request_body", description = "", required = true) @Valid @RequestBody Mono<Map<String, Object>> requestBody,
        @Size(min = 1, max = 100) @Parameter(name = "X-Tenant-ID", description = "The tenant identifier for multi-tenant isolation", in = ParameterIn.HEADER) @RequestHeader(value = "X-Tenant-ID", required = false) @Nullable String xTenantID,
        @Parameter(hidden = true) final ServerWebExchange exchange
    );

}