# libs/shared/openapi-spec/src/main/resources/openapi/bundled-synaptiq-v1.yaml
openapi: 3.1.0
info:
  title: Synaptiq API
  description: |
    AI-Powered Conversational Commerce Platform — REST API

    Synaptiq enables businesses to deploy intelligent, context-aware AI shopping assistants
    with full catalog management, conversation orchestration, workflow automation,
    and multi-tenant governance. This specification covers all bounded contexts:
    Authentication, Tenants, Catalog, Chat, Actions, Workflows, Analytics,
    Branding, Tenant Configuration, and Schema Registry.
  version: 1.0.0
  contact:
    name: Synaptiq Team
    url: https://synaptiq.dev
  license:
    name: Proprietary
    url: https://synaptiq.dev/license
servers:
  - url: http://localhost:8080
    description: Local development server
  - url: https://api.synaptiq.dev
    description: Production server
...
paths:
...
  /api/v1/catalog/schema:
    get:
      operationId: getActiveSchema
      summary: Get active catalog schema
      tags:
        - Catalog
      parameters:
        - $ref: '#/components/parameters/TenantIdHeader'
      responses:
        '200':
          description: Active schema
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CatalogSchemaResponse'
  /api/v1/catalog/items:
    post:
      operationId: createCatalogItem
      summary: Create a catalog item
      tags:
        - Catalog
      parameters:
        - $ref: '#/components/parameters/TenantIdHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCatalogItemRequest'
      responses:
        '201':
          description: Item created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CatalogItemResponse'
    get:
      operationId: listCatalogItems
      summary: List catalog items
      tags:
        - Catalog
      parameters:
        - $ref: '#/components/parameters/TenantIdHeader'
        - name: statusFilter
          in: query
          schema:
            type: string
        - name: skip
          in: query
          schema:
            type: integer
            default: 0
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Item list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CatalogItemListResponse'
  /api/v1/catalog/items/{itemId}:
    get:
      operationId: getCatalogItem
      summary: Get a catalog item
      tags:
        - Catalog
      parameters:
        - $ref: '#/components/parameters/TenantIdHeader'
        - name: itemId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Item details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CatalogItemResponse'
    patch:
      operationId: updateCatalogItem
      summary: Update a catalog item
      tags:
        - Catalog
      parameters:
        - $ref: '#/components/parameters/TenantIdHeader'
        - name: itemId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '200':
          description: Item updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CatalogItemResponse'
    delete:
      operationId: deleteCatalogItem
      summary: Delete a catalog item
      tags:
        - Catalog
      parameters:
        - $ref: '#/components/parameters/TenantIdHeader'
        - name: itemId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Item deleted
  /api/v1/catalog/import:
    post:
      operationId: importCatalogItems
      summary: Import catalog items from CSV
      tags:
        - Catalog
      parameters:
        - $ref: '#/components/parameters/TenantIdHeader'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: Import results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CatalogImportResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    TenantIdHeader:
      name: X-Tenant-ID
      in: header
      required: false
      description: The tenant identifier for multi-tenant isolation
      schema:
        type: string
        minLength: 1
        maxLength: 100
  schemas:
...
    CatalogItemStatus:
      type: string
      description: Status of a catalog item
      enum:
        - DRAFT
        - PUBLISHED
        - ARCHIVED
...
    CatalogSchemaResponse:
      type: object
      properties:
        id:
          type: string
        tenantId:
          type: string
        name:
          type: string
        fields:
          type: array
          items:
            $ref: '#/components/schemas/SchemaFieldResponse'
        schemaVersion:
          type: integer
        active:
          type: boolean
    SchemaFieldResponse:
      type: object
      properties:
        fieldId:
          type: string
        label:
          type: string
        type:
          type: string
        required:
          type: boolean
        searchable:
          type: boolean
        displayable:
          type: boolean
        filterable:
          type: boolean
        displayOrder:
          type: integer
        enumValues:
          type: array
          items:
            type: string
    CreateCatalogItemRequest:
      type: object
      required:
        - data
      properties:
        data:
          type: object
          additionalProperties: true
        status:
          $ref: '#/components/schemas/CatalogItemStatus'
    CatalogItemResponse:
      type: object
      properties:
        id:
          type: string
        tenantId:
          type: string
        status:
          $ref: '#/components/schemas/CatalogItemStatus'
        data:
          type: object
          additionalProperties: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CatalogItemListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/CatalogItemResponse'
        total:
          type: integer
          format: int64
    CatalogImportResponse:
      type: object
      properties:
        imported:
          type: integer
        skipped:
          type: integer
        errors:
          type: array
          items:
            type: string
...